chqnge

+++ Libft/Part1-libc/ft_bzero.c
@@ -14,10 +14,10 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		s[i] = 0;
 		i++;
